backend/requirements.txt
google-adk>=0.1.0
firebase-admin>=6.5.0
PyPDF2>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sendgrid>=6.11.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
watchdog>=3.0.0
python-docx>=1.1.0
